+++ src/javascript/app_2/Assets/Common/icon_close.jsx
