+++ src/javascript/app_2/Assets/Common/index.js
